Created ViewScheduleCommand and refactored some code.
ViewScheduleCommand is used to view the schedule of tasks in a certain date using 'view <date>'.
Error message for incorrect date format is also printed directly from Ui.
Refactored code and added new method in Time to make it clear if the string being read is for LocalTime or LocalDateTime.
Dates for tasks are now stored in Command rather than only in event and deadline to streamline the tasks.
In Task: getDate() will now return the LocalDateTime variable, and getDateStr will return the message to be printed.