Basic Do After functionality
When typing 'doafter' with '/itemNumber' and '/description',
the doAfterDescription will be populated. When printing a
list, this new description will appear on the line under the
main task that it is part of. Deleting tasks will delete the
whole task, and a 'doafter' command will overwrite any current
do-after tasks.