Added ModifyEntryCommand, and updated ModifyParser and DollaData, as well as other affected files.
ModifyParser: When modify is initiated, the mode changes to 'modify' and
then uses the respective modes' add parsing to verify the input.
ModifyEntryCommand: Basically acts as a 'addEntryCommand', but replaces
an already exisiting entry.
DollaData: Stores information required for modification of lists, and
now contains a method to replace a log within a loglist when in modify mode.