Revamp delete credit card bill feature
Revamp delete credit card bill feature as previous implementation was unintuitive to the user, prone to data corruption, and exploitable by user to increase his money. This new implementation is more intuitive as it no longer requires user to enter savings account expenditure and deposit number as the program will automatically look for the appropriate transactions.